(trunk) Merged changesets log:/gplates/branches/0.9.8@7378:7391 from source:/gplates/branches/0.9.8@7391 to source:/gplates/trunk@7376
This is "merging one branch to another without previous sync". The merge was performed using Tortoise SVN's Merge tool, from
source:/gplates/branches/0.9.8@7377 to
source:/gplates/branches/0.9.8@7391, 
from inside the root directory of a working copy of the trunk.

git-svn-id: http://svn.gplates.org/projects/gplates/trunk@7392 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c

+++ src/qt-widgets/AssignReconstructionPlateIdsDialog.cc
@@ -24,6 +24,8 @@
  */
 
 #include <vector>
+#include <boost/cast.hpp>
+#include <QHeaderView>
 #include <QMessageBox>
 #include <QString>
 
@@ -445,7 +447,7 @@ GPlatesQtWidgets::AssignReconstructionPlateIdsDialog::add_row(
 	// The rows in the QTableWidget and our internal file sequence should be in sync.
 	const int row = table_feature_collections->rowCount();
 	GPlatesGlobal::Assert<GPlatesGlobal::AssertionFailureException>(
-			row == d_file_state_seq.size(),
+			row == boost::numeric_cast<int>(d_file_state_seq.size()),
 			GPLATES_ASSERTION_SOURCE);
 
 	// Add a row.
